+++ templatest/_objects.py
@@ -59,11 +59,19 @@ class Registered(_MutableSequence[Template]):
         """
         return tuple(i.name for i in self)
 
+    def getbyname(self, name: str) -> _t.Optional[Template]:
+        """Get a template by name if it exists.
+
+        :param name: Name assigned to template.
+        :return: :class:`Template` object.
+        """
+        return next((i for i in self if i.name == name), None)
+
     def getindex(self, name: str) -> _t.Optional[int]:
         """Get the index of a template by name if it exists.
 
         :param name: Name assigned to template.
         :return: Index of the template if it exists, else None.
         """
-        result = next((i for i in self if i.name == name), None)
-        return result if result is None else self.index(result)
+        template = self.getbyname(name)
+        return template if template is None else self.index(template)